+++ app/build.gradle
@@ -75,6 +75,7 @@ dependencies {
 
     implementation "androidx.datastore:datastore-preferences:1.0.0"
 
+    implementation "androidx.activity:activity:1.7.0"
 
     //DAGGER
     implementation "com.google.dagger:hilt-android:2.44"
