+++ app/src/main/java/com/example/lesson001/data/Note.kt
@@ -2,6 +2,7 @@ package com.example.lesson001.data
 
 data class Note(
     val id: Long,
-    val text: String,
+    val text: String
+
 
 )
