+++ app/src/main/res/layout/item_note.xml
@@ -1,20 +1,35 @@
 <?xml version="1.0" encoding="utf-8"?>
 <com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     style="@style/Widget.Material3.CardView.Elevated"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_margin="8dp"
-    >
+    android:layout_margin="8dp">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:scaleType="centerCrop"
+        app:layout_constraintDimensionRatio="1:1"
+        app:layout_constraintStart_toStartOf="parent"
+        />
 
     <TextView
         android:id="@+id/textViewText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="16dp"
-        android:longClickable="true"
         android:textAppearance="@style/TextAppearance.Material3.BodyLarge"
-        tools:text="Note text that resizes the card vertically to fit itself. It can be very long, but let’s settle on 180 characters as the limit"
-        />
+        tools:text="Note text that resizes the card vertically to fit itself. It can be very long, but let’s settle on 180 characters as the limit" />
 
+
+    </LinearLayout>
 </com.google.android.material.card.MaterialCardView>
